Add new test files
